Boom! Static multi-tenancy per database is working.
