+++ src/Marten/Marten.csproj
@@ -53,7 +53,7 @@
         <PackageReference Include="Remotion.Linq" Version="2.2.0" />
         <PackageReference Include="System.Threading.Tasks.Dataflow" Version="6.0.0" />
         <PackageReference Include="System.Text.Json" Version="6.0.0" />
-        <PackageReference Include="Weasel.Postgresql" Version="5.1.0" />
+        <PackageReference Include="Weasel.Postgresql" Version="5.1.2" />
     </ItemGroup>
 
     <!--SourceLink specific settings-->
